requirements.txt
customtkinter==5.2.2
pillow==10.3.0
cryptography==42.0.6
python-dotenv==1.0.1